Clean up CLI options and commit generated Usage.md
Also:
- Add optional man page generation to `build.sh`
- Update CI workflow to generate a man page for `pretty-php`

+++ scripts/build.sh
@@ -17,9 +17,10 @@ function usage() {
     else
         cat
     fi <<EOF
-usage: ${0##*/}             build the current working tree
-       ${0##*/} latest      build the closest version reachable from HEAD
-       ${0##*/} v<VERSION>  build a tagged version${1:+
+usage: ${0##*/}                   build the current working tree
+       ${0##*/} man               build a man page for the current working tree
+       ${0##*/} [man] latest      build the closest version reachable from HEAD
+       ${0##*/} [man] v<VERSION>  build a tagged version${1:+
 }
 EOF
     exit
@@ -35,18 +36,26 @@ PACKAGE=$(composer show --self --name-only) &&
 REPO=$PWD
 DIST=build/dist
 PHAR=$DIST/$PACKAGE.phar
+MAN=$DIST/$PACKAGE.1
 
+BUILD_PHAR=$(($# ? 0 : 1))
+BUILD_MAN=0
 FROM_GIT=0
 VERSION=
 while (($#)); do
     case "$1" in
     latest)
+        BUILD_PHAR=1
         FROM_GIT=1
         VERSION=
         ;;
     v[0-9]*)
+        BUILD_PHAR=1
         FROM_GIT=1
         VERSION=$1
+        ;;
+    man)
+        BUILD_MAN=1
         ;;
     -h | --help)
         usage
@@ -77,31 +86,43 @@ if ((FROM_GIT)); then
         git -C "$TEMP_DIR" fetch --progress --no-recurse-submodules --depth=1 origin +refs/tags/"$VERSION":refs/tags/"$VERSION" &&
         git -C "$TEMP_DIR" -c advice.detachedHead=false checkout --progress --force refs/tags/"$VERSION" ||
         die "error checking out $PACKAGE $VERSION"
-else
+elif ((BUILD_PHAR)); then
     printf '==> copying %s working tree to %s\n' "$PACKAGE" "$TEMP_DIR"
     cp -Lpr "$REPO"/!(box|build|vendor) "$TEMP_DIR" ||
         die "error copying working tree to $TEMP_DIR"
 fi
 
-printf '==> installing %s production dependencies in %s\n' "$PACKAGE" "$TEMP_DIR"
-composer install -d "$TEMP_DIR" --no-plugins --no-interaction --no-dev
+if ((BUILD_PHAR)); then
+    printf '==> installing %s production dependencies in %s\n' "$PACKAGE" "$TEMP_DIR"
+    composer install -d "$TEMP_DIR" --no-plugins --no-interaction --no-dev
 
-printf '==> installing humbug/box in %s/box\n' "$REPO"
-composer install -d box --no-plugins --no-interaction
+    printf '==> installing humbug/box in %s/box\n' "$REPO"
+    composer install -d box --no-plugins --no-interaction
 
-printf "==> running 'box compile' in %s\\n" "$TEMP_DIR"
-{ [[ -f $TEMP_DIR/box.json ]] ||
-    [[ -f $TEMP_DIR/box.json.dist ]] ||
-    cp -v "$REPO/box.json" "$TEMP_DIR/box.json"; } &&
-    php -d phar.readonly=off box/vendor/bin/box compile -d "$TEMP_DIR" --no-interaction
+    printf "==> running 'box compile' in %s\\n" "$TEMP_DIR"
+    { [[ -f $TEMP_DIR/box.json ]] ||
+        [[ -f $TEMP_DIR/box.json.dist ]] ||
+        cp -v "$REPO/box.json" "$TEMP_DIR/box.json"; } &&
+        php -d phar.readonly=off box/vendor/bin/box compile -d "$TEMP_DIR" --no-interaction
 
-printf '==> finalising build\n'
-TEMP_PHAR=("$TEMP_DIR/$DIST"/*)
-[[ ${#TEMP_PHAR[@]} -eq 1 ]] ||
-    die "output missing or invalid"
-rm -f "$PHAR" &&
-    mkdir -pv "$DIST" &&
-    cp -pv "$TEMP_PHAR" "$PHAR" ||
-    die "error copying $TEMP_PHAR to $PHAR"
+    printf '==> finalising build\n'
+    TEMP_PHAR=("$TEMP_DIR/$DIST"/*)
+    [[ ${#TEMP_PHAR[@]} -eq 1 ]] ||
+        die "output missing or invalid"
+    rm -f "$PHAR" &&
+        mkdir -pv "$DIST" &&
+        cp -pv "$TEMP_PHAR" "$PHAR" ||
+        die "error copying $TEMP_PHAR to $PHAR"
 
-printf ' -> PHP archive created at %s/%s\n' "$REPO" "$PHAR"
+    printf ' -> PHP archive created at %s/%s\n' "$REPO" "$PHAR"
+fi
+
+if ((BUILD_MAN)); then
+    ((BUILD_PHAR)) || PHAR=bin/pretty-php
+    printf '==> generating man page for %s\n' "$PHAR"
+    rm -f "$MAN" &&
+        "$PHAR" _man | pandoc --standalone --to man -o "$MAN" ||
+        die "error creating man page at $MAN"
+
+    printf ' -> man page created at %s/%s\n' "$REPO" "$MAN"
+fi
